+++ app/src/main/java/com/tft_mvvm/app/mapper/ChampMapper.kt
@@ -2,7 +2,7 @@ package com.tft_mvvm.app.mapper
 
 import com.example.common_jvm.mapper.Mapper
 import com.tft_mvvm.app.features.champ.model.Champ
-import com.tft_mvvm.domain.features.champs.model.ChampListEntity
+import com.tft_mvvm.domain.features.model.ChampListEntity
 
 class ChampMapper : Mapper<ChampListEntity.Champ, Champ>() {
     override fun map(input: ChampListEntity.Champ): Champ {
