+++ app/src/main/java/com/tft_mvvm/app/mapper/TeamBuilderMapper.kt
@@ -1,10 +1,8 @@
 package com.tft_mvvm.app.mapper
 
-import com.example.common_jvm.extension.nullable.defaultEmpty
 import com.example.common_jvm.mapper.Mapper
 import com.tft_mvvm.app.features.champ.model.TeamBuilder
-import com.tft_mvvm.domain.features.champs.model.ChampListEntity
-import com.tft_mvvm.domain.features.champs.model.TeamBuilderListEntity
+import com.tft_mvvm.domain.features.model.TeamBuilderListEntity
 
 class TeamBuilderMapper(
     private val champMapper: ChampMapper
