+++ data/src/main/java/com/tft_mvvm/data/features/champs/mapper/ItemListMapper.kt
@@ -2,7 +2,7 @@ package com.tft_mvvm.data.features.champs.mapper
 
 import com.example.common_jvm.mapper.Mapper
 import com.tft_mvvm.data.local.model.ItemListDBO
-import com.tft_mvvm.domain.features.champs.model.ItemListEntity
+import com.tft_mvvm.domain.features.model.ItemListEntity
 
 class ItemListMapper : Mapper<ItemListDBO.ItemDBO, ItemListEntity.Item>() {
     override fun map(input: ItemListDBO.ItemDBO): ItemListEntity.Item {
