+++ data/src/main/java/com/tft_mvvm/data/features/champs/mapper/TeamListMapper.kt
@@ -2,7 +2,7 @@ package com.tft_mvvm.data.features.champs.mapper
 
 import com.example.common_jvm.mapper.Mapper
 import com.tft_mvvm.data.local.model.TeamListDBO
-import com.tft_mvvm.domain.features.champs.model.TeamListEntity
+import com.tft_mvvm.domain.features.model.TeamListEntity
 
 class TeamListMapper: Mapper<TeamListDBO.TeamDBO, TeamListEntity.Team>() {
     override fun map(input: TeamListDBO.TeamDBO): TeamListEntity.Team {
