+++ domain/src/main/java/com/tft_mvvm/domain/di/DomainModule.kt
@@ -1,6 +1,6 @@
 package com.tft_mvvm.domain.di
 
-import com.tft_mvvm.domain.features.champs.di.domainModule
+import com.tft_mvvm.domain.features.di.domainModule
 import org.koin.core.module.Module
 
 fun getDomainKoinModule(): Module {
