+++ domain/src/main/java/com/tft_mvvm/domain/features/model/ClassAndOriginListEntity.kt
@@ -1,4 +1,4 @@
-package com.tft_mvvm.domain.features.champs.model
+package com.tft_mvvm.domain.features.model
 
 data class ClassAndOriginListEntity(val classAndOrigin: List<ClassAndOrigin>) {
     data class ClassAndOrigin(
