+++ domain/src/main/java/com/tft_mvvm/domain/features/model/TeamBuilderListEntity.kt
@@ -1,8 +1,8 @@
-package com.tft_mvvm.domain.features.champs.model
+package com.tft_mvvm.domain.features.model
 
 data class TeamBuilderListEntity(val teamBuilders: List<TeamsBuilder>) {
     data class TeamsBuilder(
         val name:String,
-        val champs:ChampListEntity
+        val champs: ChampListEntity
     )
 }
