+++ domain/src/main/java/com/tft_mvvm/domain/features/repository/RepoRepository.kt
@@ -1,9 +1,11 @@
-package com.tft_mvvm.domain.features.champs.repository
+package com.tft_mvvm.domain.features.repository
 
 import com.example.common_jvm.exception.Failure
 import com.example.common_jvm.function.Either
-import com.tft_mvvm.domain.features.champs.model.*
-import java.awt.Stroke
+import com.tft_mvvm.domain.features.model.ChampListEntity
+import com.tft_mvvm.domain.features.model.ClassAndOriginListEntity
+import com.tft_mvvm.domain.features.model.ItemListEntity
+import com.tft_mvvm.domain.features.model.TeamBuilderListEntity
 
 interface RepoRepository {
     suspend fun getChamps(
