+++ domain/src/main/java/com/tft_mvvm/domain/features/usecase/GetChampsByClassUseCase.kt
@@ -1,14 +1,14 @@
-package com.tft_mvvm.domain.features.champs.usecase
+package com.tft_mvvm.domain.features.usecase
 
 import com.example.common_jvm.exception.Failure
 import com.example.common_jvm.function.Either
 import com.tft_mvvm.domain.base.usecase.UseCase
 import com.tft_mvvm.domain.base.usecase.UseCaseParams
-import com.tft_mvvm.domain.features.champs.model.ChampListEntity
-import com.tft_mvvm.domain.features.champs.repository.RepoRepository
+import com.tft_mvvm.domain.features.model.ChampListEntity
+import com.tft_mvvm.domain.features.repository.RepoRepository
 
 class GetChampsByClassUseCase(private val repoRepository: RepoRepository):
-    UseCase<GetChampsByClassUseCase.GetChampsByClassUseCaseParam, Either<Failure,ChampListEntity>>(){
+    UseCase<GetChampsByClassUseCase.GetChampsByClassUseCaseParam, Either<Failure, ChampListEntity>>(){
     override suspend fun executeInternal(params: GetChampsByClassUseCaseParam): Either<Failure, ChampListEntity> {
         return repoRepository.getChampsByClass(
             classs = params.classs
